+++ clients/admin-ui/src/features/common/Icon/StepperCircle.tsx
@@ -1,0 +1,18 @@
+import { createIcon } from "@fidesui/react";
+
+export default createIcon({
+  displayName: "StepperCircleIcon",
+  viewBox: "0 0 24 24",
+  path: (
+    <svg
+      width="24"
+      height="24"
+      viewBox="0 0 24 24"
+      fill="none"
+      xmlns="http://www.w3.org/2000/svg"
+    >
+      <circle cx="12" cy="12" r="9" stroke="#EDF2F7" strokeWidth="2" />
+      <circle cx="12" cy="12" r="4" fill="#EDF2F7" />
+    </svg>
+  ),
+});