+++ clients/admin-ui/src/features/dataset/EditCollectionOrFieldForm.tsx
@@ -1,0 +1,59 @@
+import { Box, Button, Stack } from "@fidesui/react";
+import { Form, Formik } from "formik";
+
+import { CustomSelect, CustomTextInput } from "../common/form/inputs";
+import { DATA_QUALIFIERS } from "./constants";
+import { DatasetCollection, DatasetField } from "./types";
+
+type FormValues =
+  | Pick<DatasetField, "description" | "data_qualifier" | "data_categories">
+  | Pick<
+      DatasetCollection,
+      "description" | "data_qualifier" | "data_categories"
+    >;
+
+interface Props {
+  values: FormValues;
+  onClose: () => void;
+  onSubmit: (values: FormValues) => void;
+}
+
+const EditCollectionOrFieldForm = ({ values, onClose, onSubmit }: Props) => {
+  const initialValues: FormValues = {
+    description: values.description ?? "",
+    data_qualifier: values.data_qualifier,
+    data_categories: values.data_categories,
+  };
+
+  return (
+    <Formik initialValues={initialValues} onSubmit={onSubmit}>
+      <Form>
+        <Stack>
+          <Box>
+            <CustomTextInput name="description" label="Description" />
+          </Box>
+          <Box>
+            <CustomSelect name="data_qualifier" label="Identifiability">
+              {DATA_QUALIFIERS.map((qualifier) => (
+                <option key={qualifier.key} value={qualifier.key}>
+                  {qualifier.label}
+                </option>
+              ))}
+            </CustomSelect>
+          </Box>
+          <Box>Data Categories (todo)</Box>
+          <Box>
+            <Button onClick={onClose} mr={2} size="sm" variant="outline">
+              Cancel
+            </Button>
+            <Button type="submit" colorScheme="primary" size="sm">
+              Save
+            </Button>
+          </Box>
+        </Stack>
+      </Form>
+    </Formik>
+  );
+};
+
+export default EditCollectionOrFieldForm;